Doing some more small refactors to improve clarity
