+++ tools/approval-service/internal/service/githubrelease.go
@@ -1,0 +1,129 @@
+package service
+
+import (
+	"bytes"
+	"context"
+	"fmt"
+	"log/slog"
+	"os"
+	"text/template"
+
+	"github.com/gravitational/shared-workflows/libs/github"
+	"github.com/gravitational/shared-workflows/tools/approval-service/internal/config"
+	"github.com/gravitational/teleport/api/types"
+)
+
+// GitHubReleaseService handles approvals/rejections for deployment protection reviews on workflows
+// based on the reviewed Access Requests. This is per-repo, and contains the logic to handle the
+// decision-making process for deployment protection rules.
+type GitHubReleaseService struct {
+	ghClient *github.Client
+	org      string
+	repo     string
+
+	envToRole map[string]string
+	log       *slog.Logger
+}
+
+// NewGitHubReleaseService creates a new GitHub deployment approval handler for deployment protection rules
+// in a given GitHub organization and repository.
+func NewGitHubReleaseService(ctx context.Context, cfg config.GitHubSource, log *slog.Logger) (*GitHubReleaseService, error) {
+	key, err := os.ReadFile(cfg.Authentication.App.PrivateKeyPath)
+	if err != nil {
+		return nil, fmt.Errorf("reading private key file %q: %w", cfg.Authentication.App.PrivateKeyPath, err)
+	}
+
+	client, err := github.NewForApp(ctx, cfg.Authentication.App.AppID, cfg.Authentication.App.InstallationID, key)
+	if err != nil {
+		return nil, fmt.Errorf("creating GitHub client for app: %w", err)
+	}
+
+	h := &GitHubReleaseService{
+		log:       log,
+		org:       cfg.Org,
+		repo:      cfg.Repo,
+		envToRole: make(map[string]string),
+		ghClient:  client,
+	}
+
+	for _, env := range cfg.Environments {
+		h.envToRole[env.Name] = env.TeleportRole
+	}
+
+	return h, nil
+}
+
+// TeleportRoleForEnvironment returns the Teleport role for a given environment.
+func (h *GitHubReleaseService) TeleportRoleForEnvironment(env string) (string, error) {
+	role, ok := h.envToRole[env]
+	if !ok {
+		return "", fmt.Errorf("no Teleport role configured for environment %q", env)
+	}
+	return role, nil
+}
+
+// HandleDecisionForAccessRequestReviewed processes the decision for an access request that has been reviewed.
+// It will either approve or reject the deployment protection rule based on the state of the access request
+func (h *GitHubReleaseService) HandleDecisionForAccessRequestReviewed(ctx context.Context, status types.RequestState, env string, workflowID int64) error {
+	var decision github.PendingDeploymentApprovalState
+	switch status {
+	case types.RequestState_APPROVED:
+		decision = github.PendingDeploymentApprovalStateApproved
+	default:
+		decision = github.PendingDeploymentApprovalStateRejected
+	}
+
+	if err := h.ghClient.ReviewDeploymentProtectionRule(ctx, h.org, h.repo, workflowID, decision, env, ""); err != nil {
+		return fmt.Errorf("reviewing deployment protection rule: %w", err)
+	}
+
+	h.log.Info("Handled decision for access request reviewed", "org", h.org, "repo", h.repo, "env", env, "workflow_run_id", workflowID, "decision", decision)
+	return nil
+}
+
+// GenerateAccessRequestReason generates a reason for the access request based on the workflow run information.
+// This reason will be used in the access request to provide context for the approval.
+func (h *GitHubReleaseService) GenerateAccessRequestReason(runID int64, env string) (string, error) {
+	runInfo, err := h.ghClient.GetWorkflowRunInfo(context.Background(), h.org, h.repo, runID)
+	if err != nil {
+		return "", fmt.Errorf("getting workflow run info: %w", err)
+	}
+
+	templateData := &accessRequestReasonTemplateData{
+		Organization: h.org,
+		Repository:   h.repo,
+		WorkflowName: runInfo.Name,
+		URL:          runInfo.HTMLURL,
+		Environment:  env,
+		WorkflowID:   runID,
+		Requester:    runInfo.Requester,
+	}
+
+	var buff bytes.Buffer
+	if err := reasonTmpl.Execute(&buff, templateData); err != nil {
+		return "", fmt.Errorf("executing reason template: %w", err)
+	}
+	return buff.String(), nil
+}
+
+var reasonTmpl = template.Must(template.New("").Parse(`GitHub Deployment Review for:
+
+Repository: {{ .Organization}}/{{ .Repository }}
+Workflow name: {{ .WorkflowName }}
+URL: {{ .URL }}
+Environment: {{ .Environment }}
+Workflow run ID: {{ .WorkflowID }}
+Requester: {{ .Requester }}
+
+This request was generated by the pipeline approval service.
+`))
+
+type accessRequestReasonTemplateData struct {
+	Organization string
+	Repository   string
+	WorkflowName string
+	URL          string
+	Environment  string
+	WorkflowID   int64
+	Requester    string
+}
